lesson-5

+++ src/services/user.service.ts
@@ -1,4 +1,6 @@
-import { IUser, IUserDTO } from "../interfaces/user.interface";
+import { StatusCodesEnum } from "../enums/status-codes.enum";
+import { ApiError } from "../errors/api.error";
+import { IUser, IUserCreateDTO } from "../interfaces/user.interface";
 import { userRepository } from "../repositoryes/user.repository";
 
 class UserService {
@@ -6,20 +8,41 @@ class UserService {
     return userRepository.getAll();
   }
 
-  public getById(id: string): Promise<IUser> {
-    return userRepository.getById(id);
+  public async getById(id: string): Promise<IUser> {
+    const user = await userRepository.getById(id);
+    if (!user) {
+      throw new ApiError("User not found", StatusCodesEnum.NOT_FOUND);
+    }
+    return user;
   }
 
-  public create(user: IUserDTO): Promise<IUser> {
+  public create(user: IUserCreateDTO): Promise<IUser> {
     return userRepository.create(user);
   }
 
-  public updateById(id: string, user: IUserDTO): Promise<IUser> {
-    return userRepository.updateById(id, user);
+  public async updateById(id: string, user: IUserCreateDTO): Promise<IUser> {
+    const data = await userRepository.getById(id);
+    if (!data) {
+      throw new ApiError("User not found", StatusCodesEnum.NOT_FOUND);
+    }
+    return await userRepository.updateById(id, user);
   }
 
-  public deleteById(id: string): Promise<IUser> {
-    return userRepository.deleteById(id);
+  public async deleteById(id: string): Promise<void> {
+    const user = await userRepository.getById(id);
+    if (!user) {
+      throw new ApiError("User not found", StatusCodesEnum.NOT_FOUND);
+    }
+    await userRepository.deleteById(id);
+  }
+  public async isEmailUnique(email: string): Promise<void> {
+    const user = userRepository.findByEmail(email);
+    if (user) {
+      throw new ApiError(
+        "User is already existed",
+        StatusCodesEnum.BED_REQUEST,
+      );
+    }
   }
 }
